% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.HierFabs.R
\name{predict.HierFabs}
\alias{predict.HierFabs}
\title{Model prediction based on a fitted HierFabs object.}
\usage{
\method{predict}{HierFabs}(object, new.G, new.E, allpath = FALSE, ...)
}
\arguments{
\item{object}{Fitted \code{'HierFabs'} model object.}

\item{new.G}{Matrix of new values for genes.}

\item{new.E}{An optional new values for environment variables.}

\item{allpath}{allpath = T will output all the predictions on the solution path. allpath = FALSE will only output the optimal solution selected in the 'HierFabs' object.}

\item{\dots}{Not used. Other arguments to predict.}
}
\value{
Predictions for each observation.
}
\description{
Similar to the usual predict methods, this function returns predictions from a fitted \code{'HierFabs'} object.
}
\seealso{
\code{\link{HierFabs}}
}
