% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.HierFabs.R
\name{print.HierFabs}
\alias{print.HierFabs}
\title{Result summary of a fitted HierFabs x.}
\usage{
print.HierFabs(x, digits = max(getOption("digits") - 2, 3), ...)
}
\arguments{
\item{x}{Fitted \code{'HierFabs'} model object..}

\item{digits}{The number of significant digits for the coefficient estimates.}

\item{\dots}{Not used. Other arguments to print.}
}
\value{
No value is returned.
}
\description{
Similar to the usual print methods, this function summarize results from a fitted \code{'HierFabs'} object.
}
\seealso{
\code{\link{HierFabs}}
}
